+++ base/server/src/main/java/com/netscape/cms/realm/RealmConfig.java
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
 //
-package org.dogtagpki.acme.realm;
+package com.netscape.cms.realm; 
 
 import java.util.Collection;
 import java.util.LinkedHashMap;
@@ -23,7 +23,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
  */
 @JsonInclude(Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown=true)
-public class ACMERealmConfig {
+public class RealmConfig {
 
     @JsonProperty("class")
     private String className;
@@ -69,14 +69,14 @@ public class ACMERealmConfig {
         return mapper.writeValueAsString(this);
     }
 
-    public static ACMERealmConfig fromJSON(String json) throws Exception {
+    public static RealmConfig fromJSON(String json) throws Exception {
         ObjectMapper mapper = new ObjectMapper();
-        return mapper.readValue(json, ACMERealmConfig.class);
+        return mapper.readValue(json, RealmConfig.class);
     }
 
-    public static ACMERealmConfig fromProperties(Properties props) throws Exception {
+    public static RealmConfig fromProperties(Properties props) throws Exception {
 
-        ACMERealmConfig config = new ACMERealmConfig();
+        RealmConfig config = new RealmConfig();
 
         for (Entry<Object, Object> entry : props.entrySet()) {
